% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_svm.R
\encoding{UTF-8}
\name{DT_replace_NA}
\alias{DT_replace_NA}
\title{DT_replace_NA}
\usage{
DT_replace_NA(DT)
}
\arguments{
\item{DT}{data.table which may contain NA values.}
}
\value{
DT data.table where any NAs have been converted to zero.
}
\description{
Function to replace NAs with zero in a data.table.
}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{replace NAs with zero in a data.table}
